requirements.txt
# Core dependencies
smbus2>=0.4.0
filelock>=3.12.0
PyYAML>=6.0
pydantic>=2.0
colorlog>=6.7

# Optional: gRPC support (install with: pip install -r requirements.txt -e ".[grpc]")
# grpcio>=1.60.0
# grpcio-tools>=1.60.0
# grpcio-reflection>=1.60.0
# protobuf>=4.25.0
